image/cas: Add a generic CAS interface
And implement that interface for tarballs based on the specs
image-layout.  I plan on adding other backends later, but this is
enough for a proof of concept.

Also add a new 'cas get' subcommand to oci-image-tool so folks can
access the new read functionality from the command line.

In a subsequent commit, I'll replace oci-image-tools walker
functionality with this new API.

The Context interface follows the pattern recommended in [1], allowing
callers to cancel long running actions (e.g. push/pull over the
network for engine implementations that communicate with a remote
store).

[1]: https://blog.golang.org/context

Signed-off-by: W. Trevor King <wking@tremily.us>