image/refs: Add a generic name-based reference interface
And implement that interface for tarballs based on the specs
image-layout.  I plan on adding other backends later, but this is
enough for a proof of concept.

Also add new 'refs list' and 'refs get' subcommands to oci-image-tool
so folks can access the new read functionality from the command line.

The Engine.List interface returns channels instead of a slice, which
has two main benefits:

* There's no need to allocate a slice for the results, so calls with
  large (or negative) 'size' values can be made without consuming
  large amounts of memory.

* The name collection and processing can happen concurrently, so:
  * We don't waste cycles collecting names we won't use.
  * Slow collection can happen in the background if/when the consumer
    is blocked on something else.

See [1] for more background and examples of the idioms I've tried to
mimic.

In a subsequent commit, I'll replace oci-image-tools walker
functionality with this new API.

[1]: https://blog.golang.org/pipelines

Signed-off-by: W. Trevor King <wking@tremily.us>