image/layout/tar: Add a CreateTarFile helper
The NewEngine commands for the tar-backed image-layout engines (both
the CAS and refs engines) open files O_RDWR and expect image-layout
compatible content in the tarball.  That makes sense, but for folks
who *don't* have such a tarball, a helper like CreateTarFile makes it
easy to explicitly create an empty one.

The 0666 file modes and 0777 directory modes rely on the caller's
umask to appropriately limit user/group/other permissions for the
tarball itself and any content extracted to the filesystem from the
tarball.

Expose this new functionality on the command line as:

  $ oci-image-tool init image-layout PATH

where 'init' and 'image-layout' are two separate levels in case we
support initializing additional types of repositories in the future.

Signed-off-by: W. Trevor King <wking@tremily.us>