WIP: Use local ref-engine discovery for layouts
Bump the layout to v1.1 to support this.  This makes it possible to
distribute layouts that use other protocols, for example new
ref-engine protocols or a sharded blob store [1].  You can also
reference external ref- and CAS-engines, although obviously the
utility of such depends on the availability of those external engines.

[1]: https://github.com/opencontainers/image-spec/issues/449

Signed-off-by: W. Trevor King <wking@tremily.us>